Devex: refactor file doc use cases/actions for more shared code
